+++ src/Recyclarr.Cli/Pipelines/QualitySize/PipelinePhases/QualitySizeConfigPhase.cs
@@ -10,31 +10,30 @@ public class QualitySizeConfigPhase(ILogger log, IQualitySizeGuideService guide,
 {
     public Task Execute(QualitySizePipelineContext context)
     {
-        var qualityDef = config.QualityDefinition;
-        if (qualityDef is null)
+        var configSizeData = config.QualityDefinition;
+        if (configSizeData is null)
         {
             log.Debug("{Instance} has no quality definition", config.InstanceName);
             return Task.CompletedTask;
         }
 
-        var qualityDefinitions = guide.GetQualitySizeData(config.ServiceType);
-        var selectedQuality = qualityDefinitions
-            .FirstOrDefault(x => x.Type.EqualsIgnoreCase(qualityDef.Type));
+        var guideSizeData = guide.GetQualitySizeData(config.ServiceType)
+            .FirstOrDefault(x => x.Type.EqualsIgnoreCase(configSizeData.Type));
 
-        if (selectedQuality == null)
+        if (guideSizeData == null)
         {
-            context.ConfigError = $"The specified quality definition type does not exist: {qualityDef.Type}";
+            context.ConfigError = $"The specified quality definition type does not exist: {configSizeData.Type}";
             return Task.CompletedTask;
         }
 
-        AdjustPreferredRatio(qualityDef, selectedQuality);
-        context.ConfigOutput = selectedQuality;
+        AdjustPreferredRatio(configSizeData, guideSizeData);
+        context.ConfigOutput = guideSizeData;
         return Task.CompletedTask;
     }
 
-    private void AdjustPreferredRatio(QualityDefinitionConfig qualityDefConfig, QualitySizeData selectedQuality)
+    private void AdjustPreferredRatio(QualityDefinitionConfig configSizeData, QualitySizeData guideSizeData)
     {
-        if (qualityDefConfig.PreferredRatio is null)
+        if (configSizeData.PreferredRatio is null)
         {
             return;
         }
@@ -42,20 +41,20 @@ public class QualitySizeConfigPhase(ILogger log, IQualitySizeGuideService guide,
         log.Information("Using an explicit preferred ratio which will override values from the guide");
 
         // Fix an out of range ratio and warn the user
-        if (qualityDefConfig.PreferredRatio is < 0 or > 1)
+        if (configSizeData.PreferredRatio is < 0 or > 1)
         {
-            var clampedRatio = Math.Clamp(qualityDefConfig.PreferredRatio.Value, 0, 1);
+            var clampedRatio = Math.Clamp(configSizeData.PreferredRatio.Value, 0, 1);
             log.Warning("Your `preferred_ratio` of {CurrentRatio} is out of range. " +
                 "It must be a decimal between 0.0 and 1.0. It has been clamped to {ClampedRatio}",
-                qualityDefConfig.PreferredRatio, clampedRatio);
+                configSizeData.PreferredRatio, clampedRatio);
 
-            qualityDefConfig.PreferredRatio = clampedRatio;
+            configSizeData.PreferredRatio = clampedRatio;
         }
 
         // Apply a calculated preferred size
-        foreach (var quality in selectedQuality.Qualities)
+        foreach (var quality in guideSizeData.Qualities)
         {
-            quality.Preferred = quality.InterpolatedPreferred(qualityDefConfig.PreferredRatio.Value);
+            quality.Preferred = quality.InterpolatedPreferred(configSizeData.PreferredRatio.Value);
         }
     }
 }