+++ tests/Recyclarr.Cli.Tests/Pipelines/QualitySize/PipelinePhases/QualitySizeTransactionPhaseTest.cs
@@ -18,8 +18,8 @@ public class QualitySizeTransactionPhaseTest
             {
                 Qualities = new[]
                 {
-                    new QualitySizeItem("non_existent1", 0, 2, 1),
-                    new QualitySizeItem("non_existent2", 0, 2, 1)
+                    new QualityItemWithPreferred("non_existent1", 0, 2, 1),
+                    new QualityItemWithPreferred("non_existent2", 0, 2, 1)
                 }
             },
             ApiFetchOutput = new List<ServiceQualityDefinitionItem>
@@ -46,8 +46,8 @@ public class QualitySizeTransactionPhaseTest
             {
                 Qualities = new[]
                 {
-                    new QualitySizeItem("same1", 0, 2, 1),
-                    new QualitySizeItem("same2", 0, 2, 1)
+                    new QualityItemWithPreferred("same1", 0, 2, 1),
+                    new QualityItemWithPreferred("same2", 0, 2, 1)
                 }
             },
             ApiFetchOutput = new List<ServiceQualityDefinitionItem>
@@ -84,8 +84,8 @@ public class QualitySizeTransactionPhaseTest
             {
                 Qualities = new[]
                 {
-                    new QualitySizeItem("same1", 0, 2, 1),
-                    new QualitySizeItem("different1", 0, 3, 1)
+                    new QualityItemWithPreferred("same1", 0, 2, 1),
+                    new QualityItemWithPreferred("different1", 0, 3, 1)
                 }
             },
             ApiFetchOutput = new List<ServiceQualityDefinitionItem>