+++ tests/Recyclarr.Tests/TrashGuide/QualitySize/QualityItemWithPreferredTest.cs
@@ -3,7 +3,7 @@ using Recyclarr.TrashGuide.QualitySize;
 namespace Recyclarr.Tests.TrashGuide.QualitySize;
 
 [TestFixture]
-public class QualitySizeItemTest
+public class QualityItemWithPreferredTest
 {
     private static readonly object[] PreferredTestValues =
     [
@@ -11,10 +11,13 @@ public class QualitySizeItemTest
         new object?[] {100m, 101m, true},
         new object?[] {100m, 98m, true},
         new object?[] {100m, null, true},
-        new object?[] {QualitySizeItem.PreferredUnlimitedThreshold, null, false},
-        new object?[] {QualitySizeItem.PreferredUnlimitedThreshold - 1, null, true},
+        new object?[] {QualityItemWithPreferred.PreferredUnlimitedThreshold, null, false},
+        new object?[] {QualityItemWithPreferred.PreferredUnlimitedThreshold - 1, null, true},
         new object?[]
-            {QualitySizeItem.PreferredUnlimitedThreshold, QualitySizeItem.PreferredUnlimitedThreshold, true}
+        {
+            QualityItemWithPreferred.PreferredUnlimitedThreshold, QualityItemWithPreferred.PreferredUnlimitedThreshold,
+            true
+        }
     ];
 
     [TestCaseSource(nameof(PreferredTestValues))]
@@ -23,7 +26,7 @@ public class QualitySizeItemTest
         decimal? radarrValue,
         bool isDifferent)
     {
-        var data = new QualitySizeItem("", 0, 0, guideValue);
+        var data = new QualityItemWithPreferred("", 0, 0, guideValue);
         data.IsPreferredDifferent(radarrValue)
             .Should().Be(isDifferent);
     }
@@ -34,19 +37,19 @@ public class QualitySizeItemTest
         {
             400m,
             1.0m,
-            QualitySizeItem.PreferredUnlimitedThreshold
+            QualityItemWithPreferred.PreferredUnlimitedThreshold
         },
         new[]
         {
-            QualitySizeItem.PreferredUnlimitedThreshold,
+            QualityItemWithPreferred.PreferredUnlimitedThreshold,
             1.0m,
-            QualitySizeItem.PreferredUnlimitedThreshold
+            QualityItemWithPreferred.PreferredUnlimitedThreshold
         },
         new[]
         {
-            QualitySizeItem.PreferredUnlimitedThreshold - 1m,
+            QualityItemWithPreferred.PreferredUnlimitedThreshold - 1m,
             1.0m,
-            QualitySizeItem.PreferredUnlimitedThreshold - 1m
+            QualityItemWithPreferred.PreferredUnlimitedThreshold - 1m
         },
         new[]
         {
@@ -68,54 +71,54 @@ public class QualitySizeItemTest
         decimal ratio,
         decimal expectedResult)
     {
-        var data = new QualitySizeItem("", 0, max, 0);
+        var data = new QualityItemWithPreferred("", 0, max, 0);
         data.InterpolatedPreferred(ratio).Should().Be(expectedResult);
     }
 
     [Test]
     public void AnnotatedPreferred_OutsideThreshold_EqualsSameValueWithUnlimited()
     {
-        const decimal testVal = QualitySizeItem.PreferredUnlimitedThreshold;
-        var data = new QualitySizeItem("", 0, 0, testVal);
+        const decimal testVal = QualityItemWithPreferred.PreferredUnlimitedThreshold;
+        var data = new QualityItemWithPreferred("", 0, 0, testVal);
         data.AnnotatedPreferred.Should().Be($"{testVal} (Unlimited)");
     }
 
     [Test]
     public void AnnotatedPreferred_WithinThreshold_EqualsSameStringValue()
     {
-        const decimal testVal = QualitySizeItem.PreferredUnlimitedThreshold - 1;
-        var data = new QualitySizeItem("", 0, 0, testVal);
+        const decimal testVal = QualityItemWithPreferred.PreferredUnlimitedThreshold - 1;
+        var data = new QualityItemWithPreferred("", 0, 0, testVal);
         data.AnnotatedPreferred.Should().Be($"{testVal}");
     }
 
     [Test]
     public void Preferred_AboveThreshold_EqualsSameValue()
     {
-        const decimal testVal = QualitySizeItem.PreferredUnlimitedThreshold + 1;
-        var data = new QualitySizeItem("", 0, 0, testVal);
+        const decimal testVal = QualityItemWithPreferred.PreferredUnlimitedThreshold + 1;
+        var data = new QualityItemWithPreferred("", 0, 0, testVal);
         data.Preferred.Should().Be(testVal);
     }
 
     [Test]
     public void PreferredForApi_AboveThreshold_EqualsNull()
     {
-        const decimal testVal = QualitySizeItem.PreferredUnlimitedThreshold + 1;
-        var data = new QualitySizeItem("", 0, 0, testVal);
+        const decimal testVal = QualityItemWithPreferred.PreferredUnlimitedThreshold + 1;
+        var data = new QualityItemWithPreferred("", 0, 0, testVal);
         data.PreferredForApi.Should().Be(null);
     }
 
     [Test]
     public void PreferredForApi_HighestWithinThreshold_EqualsSameValue()
     {
-        const decimal testVal = QualitySizeItem.PreferredUnlimitedThreshold - 0.1m;
-        var data = new QualitySizeItem("", 0, 0, testVal);
+        const decimal testVal = QualityItemWithPreferred.PreferredUnlimitedThreshold - 0.1m;
+        var data = new QualityItemWithPreferred("", 0, 0, testVal);
         data.PreferredForApi.Should().Be(testVal).And.Be(data.Preferred);
     }
 
     [Test]
     public void PreferredForApi_LowestWithinThreshold_EqualsSameValue()
     {
-        var data = new QualitySizeItem("", 0, 0, 0);
+        var data = new QualityItemWithPreferred("", 0, 0, 0);
         data.PreferredForApi.Should().Be(0);
     }
 }